added strategy to paddle, fixed keylistener direction, small tweaks

+++ src/main/java/pong/PongFrame.java
@@ -7,7 +7,7 @@ public class PongFrame extends JFrame {
 
     public PongFrame(PongView pongView, PaddleKeyListener paddleListener) throws HeadlessException {
         super();
-        setSize(Environment.WIDTH,Environment.HEIGHT);
+        setSize((int) Environment.WIDTH, (int) Environment.HEIGHT);
         setDefaultCloseOperation(WindowConstants.EXIT_ON_CLOSE);
         setTitle("Pong");
         setLayout(new BorderLayout());
